notifikacije fix

+++ src/main/java/org/unibl/etf/epraksa/repositories/NotificationRepository.java
@@ -7,7 +7,7 @@ import org.unibl.etf.epraksa.model.entities.Notification;
 import java.util.List;
 
 public interface NotificationRepository extends JpaRepository<Notification,Integer> {
-    List<Notification> getNotificationsByUserID(Integer userID);
+    List<Notification> getNotificationsByUserID(Long userID);
     @Query("UPDATE Notification n SET n.delivered=true WHERE n.notificationID= :notificationId")
     void setDelivered(Long notificationId);
 }
